+++ crates/sage-cloudpath/src/parquet/results.rs
@@ -411,6 +411,7 @@ pub fn serialize_matched_fragments(
                     Kind::X => "x".as_bytes().into(),
                     Kind::Y => "y".as_bytes().into(),
                     Kind::Z => "z".as_bytes().into(),
+                    Kind::ZDot => "z_dot".as_bytes().into(),
                 })
                 .collect::<Vec<ByteArray>>();